chore: port custom user render target visual test
